Update webpack dependency version
This is a temporary patch to prevent broken builds until the package is made compatible with the breaking changes introduced in webpack 2.1.0-beta.23. I'd just submit a patch to make it compatible but I haven't had time to look into what that would involve, so I'm submiting this for now.